+++ lib/job-manager.js
@@ -3,7 +3,7 @@ var Base = require('extendable-base');
 var uuid = require('uuid');
 var logger = require('log4js').getLogger('demo-job-manager');
 var events = require('backbone').Events;
-var JuttleJob = require('./juttle-job');
+var WebsocketJuttleJob = require('./ws-juttle-job');
 
 // This class handles management of jobs (running juttle programs). It
 // also brings together websocket endpoints and the outputs (sinks) of
@@ -77,7 +77,7 @@ var JobManager = Base.extend({
         logger.debug('Running program "' + options.bundle.program.slice(0, 20) + '..." with observer ' + options.observer_id);
 
         var job_id = uuid.v4();
-        var job = new JuttleJob({
+        var job = new WebsocketJuttleJob({
             job_id: job_id,
             bundle: options.bundle,
             inputs: options.inputs,
