+++ lib/juttle-job.js
@@ -4,10 +4,17 @@ var JSDP = require('juttle-jsdp');
 var JSDPValueConverter = require('./jsdp-value-converter');
 var child_process = require('child_process');
 var Promise = require('bluebird');
-var CBuffer = require('CBuffer');
 var JuttledErrors = require('./errors');
 var logger = require('log4js').getLogger('juttle-job');
 var events = require('backbone').Events;
+
+// This class handles the details of running a program from a
+// bundle, handling an explicit stop from a job manager, emitting the
+// proper events when the job has completed, etc.
+//
+// Actually passing the results of programs back to websocket and/or
+// http clients is handled by subclasses. Based on the method, the Job
+// Manager will create a WebsocketJuttleJob or HTTPJuttleJob object.
 
 var JuttleJob = Base.extend({
     initialize: function(options) {
@@ -20,63 +27,24 @@ var JuttleJob = Base.extend({
         self._inputs = options.inputs;
         self._config_path = options.config_path;
 
-        // An array of WebsocketEndpoint objects.
-        self._endpoints = options.endpoints;
-
         // A handle to the spawned child process where the juttle
         // program is actually run.
         self._child = undefined;
-
-        // If true, this job will stop running if all endpoints are
-        // removed.
-        self._run_without_endpoints = options.run_without_endpoints || false;
-
-        // A buffer of messages received from the juttle child
-        // process. These messages will be sent to new websocket
-        // connections.
-        self._saved_messages = new CBuffer(options.max_saved_messages);
 
         self._log_prefix = '(job=' + self._job_id + ') ';
 
         self._received_program_started = false;
         self._job_stopped = false;
-        self._sink_descs = undefined;
-
-        // This is used in multiple places and only depends on the job
-        // id, so create it now.
-        self._job_end_msg = {
-            type: 'job_end',
-            job_id: self._job_id
-        };
-
 
         logger.debug(self._log_prefix + 'Created job');
-    },
-
-    save: function(msg) {
-        var self = this;
-
-        self._saved_messages.push(msg);
-    },
-
-    replay: function(endpoint) {
-        var self = this;
-
-        // This works because CBuffer implements forEach.
-        endpoint.send_many(self._saved_messages);
     },
 
     describe: function() {
         var self = this;
 
-        var ep_desc = _.map(self._endpoints, function(endpoint) {
-            return endpoint.describe();
-        });
-
         return {
             job_id: self._job_id,
-            bundle: self._bundle,
-            endpoints: ep_desc
+            bundle: self._bundle
         };
     },
 
@@ -100,8 +68,7 @@ var JuttleJob = Base.extend({
             switch (msg.type) {
                 case 'data':
                     var send_data = _.extend(msg.data, {job_id: self._job_id});
-                    self.send(send_data);
-                    self.save(send_data);
+                    self.send_to_clients(send_data);
                     break;
                 case 'log':
                     var processLogger = require('log4js').getLogger(msg.name);
@@ -113,8 +80,7 @@ var JuttleJob = Base.extend({
                     break;
                 case 'warning':
                 case 'error':
-                    self.send(msg);
-                    self.save(msg);
+                    self.send_to_clients(msg);
                     break;
             }
         });
@@ -131,11 +97,16 @@ var JuttleJob = Base.extend({
             // message any longer.
             self._child = undefined;
 
-            self.job_stopped();
-
-            self._endpoints.forEach(function(endpoint) {
-                endpoint.close();
+            self.send_to_clients({
+                type: 'job_end',
+                job_id: self._job_id
             });
+
+            self._job_stopped = true;
+
+            self.events.trigger('end');
+
+            self.close();
         });
 
         self._child.on('error', function(err) {
@@ -159,7 +130,9 @@ var JuttleJob = Base.extend({
                     case 'program_started':
                         self._received_program_started = true;
                         self._sink_descs = msg.sinks;
-                        self.job_started();
+                        self.send_to_clients({type: 'job_start',
+                                             job_id: self._job_id,
+                                             sinks: self._sink_descs});
                         resolve(self._child.pid);
                         break;
                     case 'compile_error':
@@ -188,98 +161,16 @@ var JuttleJob = Base.extend({
         }
     },
 
-    add_endpoint: function(endpoint) {
-        var self = this;
-
-        logger.debug(self._log_prefix + 'Adding endpoint');
-
-        // If the associated program is running and the subprocess has
-        // sent a program_started message, send this endpoint a job
-        // started message.
-        if (self._received_program_started) {
-            logger.debug(self._log_prefix + 'Received program started, replaying messages');
-            self.job_started([endpoint]);
-            self.replay(endpoint);
-
-            // If the job has also stopped, this means that the endpoint
-            // is connecting after the program has finished. Just send the
-            // endpoint a job_stop message and close the endpoint.
-            if (self._job_stopped) {
-                logger.debug(self._log_prefix + 'Received job stopped, sending this endpoint a job_stopped message and closing');
-                endpoint.send(self._job_end_msg);
-                endpoint.close();
-                return;
-            }
-        }
-
-        self._endpoints.push(endpoint);
-
-        logger.debug(self._log_prefix + self._endpoints.length + ' endpoints remain');
-
-        // Subscribe to close events on the endpoint, removing the
-        // endpoint when closed.
-        endpoint.events.on('close', function() {
-            self.remove_endpoint(endpoint);
-        });
+    // Send a message to whoever is listening to the output of this
+    // job. This method should be overridden by subclasses.
+    send_to_clients: function(msg) {
     },
 
-    remove_endpoint: function(endpoint) {
-        var self = this;
-
-        logger.debug(self._log_prefix + 'Removing endpoint');
-
-        self._endpoints = _.without(self._endpoints, endpoint);
-
-        logger.debug(self._log_prefix + self._endpoints.length + ' endpoints remain');
-
-        if (self._endpoints.length === 0 &&
-            ! self._run_without_endpoints) {
-            // Stop the job.
-            self.stop();
-        }
-    },
-
-    // Send a job started message to the provided set of
-    // endpoints. endpoints is optional--if not provided the message
-    // is sent to all endpoints.
-    job_started: function(endpoints) {
-        var self = this;
-
-        endpoints = endpoints || self._endpoints;
-
-        logger.debug(self._log_prefix + 'Sending job_started message to ' + endpoints.length + ' endpoints');
-
-        self.send({
-            type: 'job_start',
-            job_id: self._job_id,
-            sinks: self._sink_descs
-        }, endpoints);
-    },
-
-    // Send a job stopped message to all endpoints.
-    job_stopped: function() {
-        var self = this;
-
-        logger.debug(self._log_prefix + 'Job stopped');
-
-        self.send(self._job_end_msg);
-
-        self._job_stopped = true;
-
-        self.events.trigger('end');
-    },
-
-    // Send a message to the specified set of endpoints. endpoints is
-    // optional--if not provided the message is sent to all endpoints.
-    send: function(msg, endpoints) {
-        var self = this;
-
-        endpoints = endpoints || self._endpoints;
-
-        endpoints.forEach(function(endpoint) {
-            endpoint.send(msg);
-        });
+    // Close all connections to clients. This method should be
+    // overridden by subclasses.
+    close: function() {
     }
+
 });
 
 module.exports = JuttleJob;
